The ratio calculation has been redone, so now all ratios fall between 0 and 1. A high weight has been given to time and movements, while comparisons have a lower weight in determining the ratio.

+++ src/Sort.java
@@ -1,7 +1,7 @@
 public class Sort
 {
-	static private int comparisons;
-	static private int movements;
+	static private long comparisons;
+	static private long movements;
 	static private long time;
 	static private long start;
 	static private long end;
